Move ane export TrueHashPlugin.
