+++ webpack/webpack5-localization-plugin/src/index.ts
@@ -2,6 +2,7 @@
 // See LICENSE in the project root for license information.
 
 export { LocalizationPlugin, type IStringPlaceholder as _IStringPlaceholder } from './LocalizationPlugin';
+export { TrueHashPlugin, type ITrueHashPluginOptions } from './TrueHashPlugin';
 
 export {
   IDefaultLocaleOptions,
@@ -20,5 +21,4 @@ export {
   IPseudolocalesOptions,
   IResolvedMissingTranslations
 } from './interfaces';
-
 export { ILocalizedWebpackChunk } from './webpackInterfaces';
